made the test cases pass and moved configuration module to a separate file.
